refactor: Move from application/ to application/firebase/
